:recycle: GoogleController로 분해
