+++ src/main/java/Idea/Archive/IdeaArchive/global/security/auth/MemberDetails.java
@@ -29,7 +29,7 @@ public class MemberDetails implements UserDetails {
 
     @Override
     public String getUsername() {
-        return member.getName();
+        return member.getEmail();
     }
 
     @Override
